Fixed RSEM indexing issue
The original used all of the outputs as the input for the `rsem-prepare-reference` command which resulted in errors because of too many inputs to the command (`Invalid number of arguments!`). I modified it to use the base of all of the outputs as an input to the `rsem-prepare-function` and left the output variable values as is. The result was a successful run!